Chat database adapter uses Long instead of Int

+++ src/main/kotlin/com/trollingcont/fullerene/server/repository/adapter/ChatDatabaseAdapter.kt
@@ -35,7 +35,7 @@ class ChatDatabaseAdapter(
         return autoIncrementId
     }
 
-    fun createChat(creator: String, updateTime: LocalDateTime): Int =
+    fun createChat(creator: String, updateTime: LocalDateTime): Long =
         transaction(db) {
             ChatData.insert {
                 it[ChatData.timeUpdated] = updateTime
@@ -43,14 +43,14 @@ class ChatDatabaseAdapter(
             }[ChatData.id].value
         }
 
-    fun setChatUpdateTime(chatId: Int, updateTime: LocalDateTime): Int =
+    fun setChatUpdateTime(chatId: Long, updateTime: LocalDateTime): Long =
         transaction(db) {
             ChatData.update({ ChatData.id eq chatId }) {
                 it[ChatData.timeUpdated] = updateTime
-            }
+            }.toLong()
         }
 
-    fun getChats(): List<Pair<Int, String>> =
+    fun getChats(): List<Pair<Long, String>> =
         transaction(db) {
             ChatParticipants.selectAll().map {
                 Pair(
@@ -60,7 +60,7 @@ class ChatDatabaseAdapter(
             }
         }
 
-    fun isChatParticipant(chatId: Int, username: String): Boolean =
+    fun isChatParticipant(chatId: Long, username: String): Boolean =
         transaction(db) {
             ChatParticipants.select {
                 (ChatParticipants.chatId eq chatId) and
@@ -68,7 +68,7 @@ class ChatDatabaseAdapter(
             }.count() != 0L
         }
 
-    fun getChatParticipants(chatId: Int): List<String> =
+    fun getChatParticipants(chatId: Long): List<String> =
         transaction(db) {
             ChatParticipants.select {
                 ChatParticipants.chatId eq chatId
@@ -77,7 +77,7 @@ class ChatDatabaseAdapter(
             }
         }
 
-    fun getUserChats(username: String, startPoint: Long = SELECT_ALL, count: Int = 1): List<Int> =
+    fun getUserChats(username: String, startPoint: Long = SELECT_ALL, count: Long = 1): List<Long> =
         transaction(db) {
             val query = ChatParticipants.join(ChatData, JoinType.INNER)
                 .slice(ChatParticipants.chatId)
@@ -94,21 +94,21 @@ class ChatDatabaseAdapter(
                 SELECT_LAST -> {
                     query
                         .orderBy(ChatData.timeUpdated, SortOrder.DESC)
-                        .limit(count)
+                        .limit(count.toInt())
                         .reversed()
                 }
 
                 else -> {
                     query
                         .orderBy(ChatData.timeUpdated, SortOrder.ASC)
-                        .limit(count, startPoint)
+                        .limit(count.toInt(), startPoint)
                 }
             }.map {
                 it[ChatParticipants.chatId]
             }
         }
 
-    fun addChatParticipant(chatId: Int, username: String) {
+    fun addChatParticipant(chatId: Long, username: String) {
         transaction(db) {
             ChatParticipants.insert {
                 it[ChatParticipants.chatId] = chatId
@@ -117,11 +117,11 @@ class ChatDatabaseAdapter(
         }
     }
 
-    fun deleteChatParticipant(chatId: Int, username: String): Int =
+    fun deleteChatParticipant(chatId: Long, username: String): Long =
         transaction(db) {
             ChatParticipants.deleteWhere {
                 (ChatParticipants.chatId eq chatId) and
                         (ChatParticipants.user eq username)
-            }
+            }.toLong()
         }
 }
